requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
langchain
langchain-community
langchain-groq
langchain-text-splitters
chromadb
python-dotenv